+++ tests/hints.rs
@@ -11,11 +11,11 @@ use cairo_vm::hint_processor::builtin_hint_processor::hint_utils::*;
 use common::utils::{check_output_vs_python, deprecated_cairo_python_run};
 use common::{load_input, load_output};
 use rstest::{fixture, rstest};
-use snos::hints::block_context::get_block_mapping;
-use snos::hints::hints_raw::*;
+use snos::hints::block_context::*;
+use snos::hints::execution::*;
 use snos::hints::{
-    enter_syscall_scopes, initialize_class_hashes, initialize_state_changes, load_next_tx, starknet_os_input,
-    STARKNET_OS_INPUT,
+    initialize_class_hashes, initialize_state_changes, starknet_os_input, INITIALIZE_CLASS_HASHES,
+    INITIALIZE_STATE_CHANGES, STARKNET_OS_INPUT,
 };
 use snos::io::input::StarknetOsInput;
 use snos::io::output::StarknetOsOutput;
